# core/src/main/resources/swagger.yaml
...
swagger: '2.0'
info:
  title: Apache MetaModel Membrane
  description: Data Federation as a RESTful service.
  version: "1.0"
  contact:
    name: Apache MetaModel
    url: http://metamodel.apache.org
  license:
    name: Apache License, version 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /{tenant}/{datasource}/q:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
    get:
      description: Executes a query on the datasource
      parameters:
      - name: sql
        in: query
        type: string
        description: The MetaModel-flavoured SQL query to execute
        required: true
      - name: offset
        in: query
        type: string
        description: An offset / first-row flag to set on the query
        required: false
      - name: limit
        in: query
        type: string
        description: A limit / max-rows flag to set on the query
        required: false
      responses:
        200:
          description: Query executed
          schema:
            $ref: "#/definitions/queryResponse"
        400:
          description: Failure while parsing query
          schema:
            $ref: "#/definitions/error"
        404:
          description: Datasource not found
          schema:
            $ref: "#/definitions/error"
        500:
          description: Failure while executing query
          schema:
            $ref: "#/definitions/error"
  /{tenant}/{datasource}/s/{schema}:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
    get:
      responses:
        200:
          description: Schema found
          schema:
            $ref: "#/definitions/getSchemaResponse"
        404:
          description: Schema not found
          schema:
            $ref: "#/definitions/error"
  /{tenant}/{datasource}/s/{schema}/t/{table}:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
      - name: table
        in: path
        type: string
        description: The table name
        required: true
    get:
      responses:
        200:
          description: Table found
          schema:
            $ref: "#/definitions/getTableResponse"
        404:
          description: Table not found
          schema:
            $ref: "#/definitions/error"
...
  /{tenant}/{datasource}/s/{schema}/t/{table}/c/{column}:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
      - name: table
        in: path
        type: string
        description: The table name
        required: true
      - name: column
        in: path
        type: string
        description: The column name
        required: true
    get:
      description: Gets information about a column
      responses:
        200:
          description: Query executed
          schema:
            $ref: "#/definitions/getColumnResponse"
        404:
          description: Column not found
          schema:
            $ref: "#/definitions/error"
definitions:
  queryResponse:
    description: Represents the result of a query - a dataset
    type: object
    properties:
      type:
        type: string
        description: The type of entity (dataset)
      headers:
        type: array
        description: The dataset header names
        items:
          type: string
      data:
        type: array
        description: The actual data returned by the query
        items:
          type: array
          items:
            type: object
...
  getSchemaResponse:
    type: object
    properties:
      type:
        type: string
        description: The type of entity (schema)
      name:
        type: string
        description: The schema name
      datasource:
        type: string
        description: The datasource name
      tenant:
        type: string
        description: The tenant name
      tables:
        type: array
        description: The names of the schema's tables
        items:
          type: object
          properties:
            name:
              type: string
              description: The table name
            uri:
              type: string
              description: A link to the table information
              format: uri
  getTableResponse:
    type: object
    properties:
      type:
        type: string
        description: The type of entity (table)
      name:
        type: string
        description: The table name
      schema:
        type: string
        description: The schema name
      datasource:
        type: string
        description: The datasource name
      tenant:
        type: string
        description: The tenant name
      columns:
        type: array
        description: The names of the table's columns
        items:
          type: object
          properties:
            name:
              type: string
              description: The column name
            uri:
              type: string
              description: A link to the column information
              format: uri
  getColumnResponse:
    type: object
    properties:
      type:
        type: string
        description: The type of entity (column)
      name:
        type: string
        description: The column name
      table:
        type: string
        description: The table name
      schema:
        type: string
        description: The schema name
      datasource:
        type: string
        description: The datasource name
      tenant:
        type: string
        description: The tenant name
      metadata:
        type: object
        description: Metadata about the column
        properties:
          number:
            type: integer
            description: The column number (0-based)
          size:
            type: integer
            description: The column size
          nullable:
            type: boolean
            description: Is the column nullable?
          primary-key:
            type: boolean
            description: Is the column a primary key?
          indexed:
            type: boolean
            description: Is the column indexed?
          column-type:
            type: string
            description: The column type (as interpreted/adapted by Apache MetaModel)
          native-type:
            type: string
            description: The native column type (as defined by the datasource itself)
          remarks:
            type: string
            description: Any remarks on the column
  error:
    description: Elaborates the error that occurred
    type: object
    properties:
      code:
        type: integer
        description: The HTTP status code
      message:
        type: string
        description: A humanly readable error message
      additional_details:
        type: object
        description: Any auxilary details to further elaborate the error